[refactor] #12 : string 추출, implementation 주석 추가
