+++ app/src/main/java/com/example/findu/presentation/ui/search/SearchReportFragment.kt
@@ -122,7 +122,7 @@ class SearchReportFragment : Fragment() {
                 binding.rvSearchHorizontalContent.removeItemDecorationAt(0)
             }
 
-            binding.rvSearchHorizontalContent.addItemDecoration(SpacingItemDecoration(10))
+            binding.rvSearchHorizontalContent.addItemDecoration(SearchSpacingItemDecoration(10))
             binding.rvSearchHorizontalContent.layoutManager = GridLayoutManager(requireContext(), 2)
             rvAdapter.setGridMode(true)
         } else {
