+++ app/src/main/java/com/example/findu/presentation/ui/search/SearchRescueFragment.kt
@@ -123,7 +123,7 @@ class SearchRescueFragment : Fragment() {
                 binding.rvSearchHorizontalContent.removeItemDecorationAt(0)
             }
 
-            binding.rvSearchHorizontalContent.addItemDecoration(SpacingItemDecoration(10))
+            binding.rvSearchHorizontalContent.addItemDecoration(SearchSpacingItemDecoration(10))
             binding.rvSearchHorizontalContent.layoutManager = GridLayoutManager(requireContext(), 2)
             rvAdapter.setGridMode(true)
         } else {
