+++ app/src/main/res/layout/fragment_search.xml
@@ -8,7 +8,7 @@
 
     <TextView
         android:id="@+id/tv_search_title"
-        style="@style/TextAppearance.FindU.Head3_SB_18"
+        android:textAppearance="@style/TextAppearance.FindU.Head3_SB_18"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:gravity="center"
